# openapi.yaml
openapi: 3.1.0
info:
  title: Teamovia API
  version: 1.0.0
  description: |
    API backend de la plateforme Teamovia — agent support client, agent ventes et orchestrateur.

    ## Authentification
    Toutes les routes requièrent un header `Authorization: Bearer <supabase_jwt>`.
    Le JWT contient le `workspace_id` du tenant courant.
    Les routes publiques (widget support) utilisent un `x-workspace-token` à la place.

    ## Multi-tenant
    L'isolation des données est garantie par RLS côté Supabase.
    Le backend valide le `workspace_id` extrait du JWT sur chaque requête.

  contact:
    name: Teamovia
    url: https://teamovia.com
servers:
  - url: https://api.teamovia.com/v1
    description: Production
  - url: https://api-staging.teamovia.com/v1
    description: Staging
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Supabase contenant workspace_id
    WorkspaceToken:
      type: apiKey
      in: header
      name: x-workspace-token
      description: Token public pour le widget support embarquable
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          example: Unauthorized
        code:
          type: string
          example: AUTH_REQUIRED
        details:
          type: object
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        has_more:
          type: boolean

    # ── Auth ──
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          enum: [starter, pro, enterprise]
        created_at:
          type: string
          format: date-time

    # ── Support ──
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        agent_id:
          type: string
          format: uuid
        user_ref:
          type: string
          description: Identifiant externe de l'utilisateur final (email, id, anonymous)
        channel:
          type: string
          enum: [chat, email, form]
        status:
          type: string
          enum: [open, resolved, escalated, transferred]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
    MessageResponse:
      type: object
      required: [message, session_id]
      properties:
        message:
          $ref: '#/components/schemas/Message'
        session_id:
          type: string
          format: uuid
        action_triggered:
          type: string
          nullable: true
          enum: [HANDOFF_TO_SALES, ESCALATE_TO_HUMAN, null]
          description: Action backend déclenchée automatiquement suite à la réponse
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────
paths:
...
  /workspace:
    get:
      tags: [Auth]
      summary: Récupère le workspace courant
      description: Retourne le workspace associé au JWT de l'utilisateur connecté.
      responses:
        '200':
          description: Workspace courant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /agents/support/sessions:
    get:
      tags: [Support]
      summary: Liste les sessions de support
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, escalated, transferred]
        - name: channel
          in: query
          schema:
            type: string
            enum: [chat, email, form]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Liste paginée des sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Support]
      summary: Crée une nouvelle session de support
      description: |
        Utilisé par le widget embarquable pour initier une conversation.
        Accepte `x-workspace-token` à la place du JWT pour les appels publics.
      security:
        - BearerAuth: []
        - WorkspaceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_ref:
                  type: string
                  description: Email, ID ou identifiant anonyme de l'utilisateur final
                channel:
                  type: string
                  enum: [chat, email, form]
                  default: chat
                metadata:
                  type: object
                  description: Données contextuelles optionnelles (page d'origine, UTM, etc.)
      responses:
        '201':
          description: Session créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /agents/support/sessions/{session_id}:
    get:
      tags: [Support]
      summary: Détail d'une session avec ses messages
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session et historique des messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    $ref: '#/components/schemas/Session'
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Support]
      summary: Met à jour le statut d'une session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [open, resolved, escalated, transferred]
      responses:
        '200':
          description: Session mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'

  # ══════════════════════════════════════
  # AGENT SUPPORT — MESSAGES
  # ══════════════════════════════════════
  /agents/support/sessions/{session_id}/message:
    post:
      tags: [Support]
      summary: Envoie un message et obtient la réponse de l'agent
      description: |
        Route principale du widget support.
        Le backend :
        1. Récupère le contexte de session et la mémoire client
        2. Interroge la base de connaissances (RAG)
        3. Appelle le LLM avec le prompt injecté
        4. Enregistre la réponse et les logs
        5. Déclenche les actions nécessaires (handoff, escalade)
      security:
        - BearerAuth: []
        - WorkspaceToken: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 4000
                metadata:
                  type: object
                  description: Données optionnelles (device, localisation, etc.)
      responses:
        '200':
          description: Réponse de l'agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ══════════════════════════════════════
  # AGENT SUPPORT — ESCALADE
  # ══════════════════════════════════════
...
components:
  responses:
    Unauthorized:
      description: Token manquant ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            code: AUTH_REQUIRED
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not found
            code: NOT_FOUND
    ValidationError:
      description: Données invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation error
            code: INVALID_INPUT
            details:
              content: Required field missing